HW-08: Corrected base to work with sensors

+++ gocourse08/sensor/breath.go
@@ -1,7 +1,9 @@
 package sensor
 
+import "math/rand/v2"
+
 type BreathSensor struct{}
 
 func (s *BreathSensor) ReadData() int {
-	return rand.Float64() * 10
+	return rand.Int() * 10
 }
